Last commit before refactoring of ordered savings

+++ Heuristics/src/headers/ClarkeWright.h
@@ -21,7 +21,7 @@ class ClarkeWright
     {
        int node1;
        int node2;
-       int saving;
+       float saving;
     };
     struct route_data
     {
@@ -32,7 +32,7 @@ class ClarkeWright
 	static void solve(ProblemData problemData);
 
 	private:
-	static int popElement(vector<int>* orderedSavings);
+	static float popElement(vector<float>* orderedSavings);
 	static int checkForRoute(map<int, route_data*> routes, int node1, int node2);
 	static void dump(map<int, route_data*> routes, ProblemData problemData, bool onlyFO);
